+++ src/index.ts
@@ -3,19 +3,19 @@ import type { AcornNode as AcornNode2 } from 'rollup'
 export type AcornNode<T = any> = AcornNode2 & Record<string, T>
 import type { Plugin, ResolvedConfig } from 'vite'
 import fastGlob from 'fast-glob'
+import { DEFAULT_EXTENSIONS } from 'vite-plugin-utils/constant'
+import {
+  MagicString,
+  walk,
+  relativeify,
+} from 'vite-plugin-utils/function'
 
 import {
-  JS_EXTENSIONS,
-  KNOWN_SFC_EXTENSIONS,
-  MagicString,
-  cleanUrl,
   hasDynamicImport,
   normallyImporteeRE,
-  simpleWalk,
   viteIgnoreRE,
   mappingPath,
   toLooseGlob,
-  relativeify
 } from './utils'
 import { type Resolved, Resolve } from './resolve'
 import { dynamicImportToGlob } from './dynamic-import-to-glob'
@@ -60,32 +60,30 @@ export interface Options {
 const PLUGIN_NAME = 'vite-plugin-dynamic-import'
 
 export default function dynamicImport(options: Options = {}): Plugin {
-  const extensions = JS_EXTENSIONS.concat(KNOWN_SFC_EXTENSIONS)
-  let globExtensions: string[]
   let config: ResolvedConfig
   let resolve: Resolve
+  let extensions = DEFAULT_EXTENSIONS
 
   return {
     name: PLUGIN_NAME,
     configResolved(_config) {
       config = _config
-      globExtensions = config.resolve?.extensions || extensions
       resolve = new Resolve(_config)
+      // https://github.com/vitejs/vite/blob/37ac91e5f680aea56ce5ca15ce1291adc3cbe05e/packages/vite/src/node/plugins/resolve.ts#L450
+      if (config.resolve?.extensions) extensions = config.resolve.extensions
     },
-    async transform(code, importer) {
-      const pureId = cleanUrl(importer)
-
-      if (/node_modules\/(?!\.vite\/)/.test(pureId)) return
-      if (!extensions.includes(path.extname(pureId))) return
+    async transform(code, id) {
+      if (/node_modules\/(?!\.vite\/)/.test(id)) return
+      if (!extensions.includes(path.extname(id))) return
       if (!hasDynamicImport(code)) return
-      if (options.filter?.(pureId) === false) return
+      if (options.filter?.(id) === false) return
 
       const ast = this.parse(code)
       const ms = new MagicString(code)
       let dynamicImportIndex = 0
       const runtimeFunctions: string[] = []
 
-      await simpleWalk(ast, {
+      await walk(ast, {
         async ImportExpression(node: AcornNode) {
           const importStatement = code.slice(node.start, node.end)
           const importeeRaw = code.slice(node.source.start, node.source.end)
@@ -94,7 +92,7 @@ export default function dynamicImport(options: Options = {}): Plugin {
           if (viteIgnoreRE.test(importStatement)) return
 
           // the user explicitly ignore this import
-          if (options.viteIgnore?.(importeeRaw, pureId)) {
+          if (options.viteIgnore?.(importeeRaw, id)) {
             ms.overwrite(node.source.start, node.source.start, '/*@vite-ignore*/') // append left
             return
           }
@@ -106,7 +104,7 @@ export default function dynamicImport(options: Options = {}): Plugin {
             // normally importee
             if (normallyImporteeRE.test(importee)) return
 
-            const rsld = await resolve.tryResolve(importee, importer)
+            const rsld = await resolve.tryResolve(importee, id)
             // alias or bare
             if (rsld && normallyImporteeRE.test(rsld.import.resolved)) {
               ms.overwrite(node.start, node.end, `import("${rsld.import.resolved}")`)
@@ -117,18 +115,18 @@ export default function dynamicImport(options: Options = {}): Plugin {
           const globResult = await globFiles(
             node,
             code,
-            importer,
+            id,
             resolve,
-            globExtensions,
+            extensions,
             options.loose !== false,
           )
           if (!globResult) return
 
           let { files, resolved, normally } = globResult
           // skip itself
-          files = files.filter(f => path.join(path.dirname(importer), f) !== importer)
+          files = files.filter(f => path.join(path.dirname(id), f) !== id)
           // execute the Options.onFiles
-          options.onFiles && (files = options.onFiles(files, importer) || files)
+          options.onFiles && (files = options.onFiles(files, id) || files)
 
           if (normally) {
             // normally importee (🚧-③ After `expressiontoglob()` processing)