+++ src/utils.ts
@@ -1,5 +1,9 @@
 import path from 'path'
 import type { AcornNode as AcornNode2 } from 'rollup'
+import {
+  singlelineCommentsRE,
+  multilineCommentsRE,
+} from 'vite-plugin-utils/constant'
 export type AcornNode<T = any> = AcornNode2 & Record<string, T>
 
 // ------------------------------------------------- RegExp
@@ -8,33 +12,10 @@ export const dynamicImportRE = /\bimport[\s\r\n]*?\(/
 // this is probably less accurate
 export const normallyImporteeRE = /^\.{1,2}\/[.-/\w]+(\.\w+)$/
 export const viteIgnoreRE = /\/\*\s*@vite-ignore\s*\*\//
-export const multilineCommentsRE = /\/\*(.|[\r\n])*?\*\//g
-export const singlelineCommentsRE = /\/\/.*/g
-export const queryRE = /\?.*$/s
-export const hashRE = /#.*$/s
 export const bareImportRE = /^[\w@](?!.*:\/\/)/
 export const deepImportRE = /^([^@][^/]*)\/|^(@[^/]+\/[^/]+)\//
 
-// ------------------------------------------------- const
-
-export const JS_EXTENSIONS = [
-  '.mjs',
-  '.js',
-  '.ts',
-  '.jsx',
-  '.tsx',
-  '.cjs'
-]
-export const KNOWN_SFC_EXTENSIONS = [
-  '.vue',
-  '.svelte',
-]
-
 // ------------------------------------------------- function
-
-export function cleanUrl(url: string): string {
-  return url.replace(hashRE, '').replace(queryRE, '')
-}
 
 export function hasDynamicImport(code: string) {
   code = code
@@ -43,115 +24,40 @@ export function hasDynamicImport(code: string) {
   return dynamicImportRE.test(code)
 }
 
-export async function simpleWalk(
-  ast: AcornNode,
-  visitors: {
-    [type: string]: (node: AcornNode) => void | Promise<void>,
-  }) {
-  if (!ast) return
-
-  if (Array.isArray(ast)) {
-    for (const element of ast as AcornNode[]) {
-      await simpleWalk(element, visitors)
-    }
-  } else {
-    for (const key of Object.keys(ast)) {
-      await (typeof ast[key] === 'object' && simpleWalk(ast[key], visitors))
-    }
-  }
-
-  await visitors[ast.type]?.(ast)
-}
-
-export class MagicString {
-  private overwrites: { loc: [number, number]; content: string }[]
-  private starts = ''
-  private ends = ''
-
-  constructor(
-    public str: string
-  ) { }
-
-  public append(content: string) {
-    this.ends += content
-    return this
-  }
-
-  public prepend(content: string) {
-    this.starts = content + this.starts
-    return this
-  }
-
-  public overwrite(start: number, end: number, content: string) {
-    if (end < start) {
-      throw new Error(`"end" con't be less than "start".`)
-    }
-    if (!this.overwrites) {
-      this.overwrites = []
-    }
-    this.overwrites.push({ loc: [start, end], content })
-    return this
-  }
-
-  public toString() {
-    let str = this.str
-    if (this.overwrites) {
-      const arr = [...this.overwrites].sort((a, b) => b.loc[0] - a.loc[0])
-      for (const { loc: [start, end], content } of arr) {
-        // TODO: check start or end overlap
-        str = str.slice(0, start) + content + str.slice(end)
-      }
-    }
-    return this.starts + str + this.ends
-  }
-}
-
-/**
- * @deprecated
- */
-// In some cases, glob may not be available
-// e.g. (fill necessary slash)
-//   foo* -> foo/*
-//   foo*.js -> foo/*.js
-export function tryFixGlobSlash(glob: string): string {
-  return glob.replace(/(?<![\*\/])(\*)/g, '/$1')
-}
-
-/**
- * @deprecated
- */
-// Match as far as possible
-// e.g.
-//   foo/* -> foo/**/*
-//   foo/*.js -> foo/**/*.js
-export function toDepthGlob(glob: string): string {
-  return glob.replace(/^(.*)\/\*(?!\*)/, '$1/**/*')
-}
-
 /**
  * Unlimit depth match
+ * @todo Use RegExp refactor
  */
 export function toLooseGlob(glob: string): string | string[] {
   if (glob.includes('**')) return glob
 
-  const ext = path.extname(glob)
-  if (ext) {
-    glob = glob.replace(ext, '')
+  const lastIndex = glob.lastIndexOf('*')
+  let tail = ''
+
+  if (lastIndex > -1) {
+    // foo*.js     -> glob=foo*  | tail=.js
+    // foo/*.js    -> glob=foo/* | tail=.js
+    // foo*bar.js  -> glob=foo*  | tail=bar.js
+    // foo*/bar.js -> glob=foo*  | tail=/bar.js
+    tail = glob.slice(lastIndex + 1)
+    glob = glob.slice(0, lastIndex + 1)
   }
 
   if (glob.endsWith('/*')) {
-    // foo/* -> foo/**/*
+    // foo/*    -> foo/**/*
     // foo/*.js -> foo/**/*.js
-    return glob + '*/*' + ext
+    return glob + '*/*' + tail
   }
 
   if (glob.endsWith('*')) {
-    // foo* -> [foo*, foo*/**/*]
-    // foo*.js -> [foo*.js, foo*/**/*.js]
-    return [glob + ext, glob + '/**/*' + ext]
+    // foo*        -> [foo*, foo*/**/*]
+    // foo*.js     -> [foo*.js, foo*/**/*.js]
+    // foo*bar.js  -> [foo*bar.js, foo*/**/bar.js]
+    // foo*/bar.js -> [foo*/bar.js, foo*/**/bar.js]
+    return [glob + tail, glob + '/**' + (tail.startsWith('/') ? tail : '/*' + tail)] // 🚨 No strict
   }
 
-  return glob + ext
+  return glob + tail
 }
 
 /**
@@ -189,17 +95,4 @@ export function mappingPath(paths: string[], alias?: Record<string, string>) {
   }
 
   return maps
-}
-
-/**
- * 🚧-②
- */
- export function relativeify(relative: string) {
-  if (relative === '') {
-    return '.'
-  }
-  if (!relative.startsWith('.')) {
-    return './' + relative
-  }
-  return relative
 }