+++ CHANGELOG.md
@@ -1,10 +1,12 @@
 
 ## [2022-09-03] v1.2.1
 
-- 5940444 feat(🌱): support "type": "module"
+- 5940444 feat(🌱): support "type": "module" #29
 
-## [2022-08-29] v1.2.0
+## [2022-08-30] v1.2.0
 
+- 8941f37 fix(🐞): nested try error #27
+- b0be8fa refactor: use vite-plugin-utils instead utils.ts
 - 03dd17e refactor: mappingPath update params. alias instead resolved
 - 5bc8537 feat: add relativeify
 - 8ecf926 feat: add findString, relative into Resolved